Removal of obsolete operators
