+++ library/src/main/java/com/datatorrent/lib/io/IdempotentStorageManager.java
@@ -36,7 +36,7 @@ import com.datatorrent.api.DAG;
 import com.datatorrent.api.StorageAgent;
 import com.datatorrent.api.annotation.Stateless;
 
-import com.datatorrent.lib.io.fs.AbstractFSDirectoryInputOperator;
+import com.datatorrent.lib.io.fs.AbstractFileInputOperator;
 import com.datatorrent.lib.util.FSStorageAgent;
 
 /**
@@ -60,7 +60,7 @@ public interface IdempotentStorageManager extends StorageAgent, Component<Contex
   /**
    * When an operator can partition itself dynamically then there is no guarantee that an input state which was being handled
    * by one instance previously will be handled by the same instance after partitioning. <br/>
-   * For eg. An {@link AbstractFSDirectoryInputOperator} instance which reads a File X till offset l (not check-pointed) may no longer be the
+   * For eg. An {@link AbstractFileInputOperator} instance which reads a File X till offset l (not check-pointed) may no longer be the
    * instance that handles file X after repartitioning as no. of instances may have changed and file X is re-hashed to another instance. <br/>
    * The new instance wouldn't know from what point to read the File X unless it reads the idempotent storage of all the operators for the window
    * being replayed and fix it's state.
