+++ library/src/main/java/com/datatorrent/lib/io/fs/AbstractFileOutputOperator.java
@@ -83,9 +83,9 @@ import com.datatorrent.lib.counters.BasicCounters;
  * @param <INPUT> This is the input tuple type.
  */
 @OperatorAnnotation(checkpointableWithinAppWindow = false)
-public abstract class AbstractFSWriter<INPUT> extends BaseOperator
+public abstract class AbstractFileOutputOperator<INPUT> extends BaseOperator
 {
-  private static final Logger LOG = LoggerFactory.getLogger(AbstractFSWriter.class);
+  private static final Logger LOG = LoggerFactory.getLogger(AbstractFileOutputOperator.class);
 
   private static final String TMP_EXTENSION = ".tmp";
 
@@ -192,7 +192,7 @@ public abstract class AbstractFSWriter<INPUT> extends BaseOperator
     }
   };
 
-  public AbstractFSWriter()
+  public AbstractFileOutputOperator()
   {
     endOffsets = Maps.newHashMap();
     counts = Maps.newHashMap();
